controller: Implement Homepage Controller and bind it react
