+++ app/helpers/homepage_helper.rb
@@ -1,0 +1,2 @@
+module HomepageHelper
+end
