+++ app/javascript/packs/index.jsx
@@ -1,0 +1,33 @@
+/* Run this example by adding <%= javascript_pack_tag 'index' %>
+to the head of your layout file,
+like app/views/layouts/application.html.erb.
+All it does is render <div>Hello React</div> at the bottom
+of the page. */
+
+import React from 'react';
+import ReactDOM from 'react-dom';
+import PropTypes from 'prop-types';
+
+const Hello = props => (
+  <div>
+    Hello
+    {' '}
+    {props.name}
+    !
+  </div>
+);
+
+Hello.defaultProps = {
+  name: 'Alex',
+};
+
+Hello.propTypes = {
+  name: PropTypes.string,
+};
+
+document.addEventListener('DOMContentLoaded', () => {
+  ReactDOM.render(
+    <Hello name="React" />,
+    document.body.appendChild(document.createElement('div')),
+  );
+});
