+++ config/routes.rb
@@ -1,3 +1,4 @@
 Rails.application.routes.draw do
+  root 'homepage#index'
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
 end
